Some progress on Configs

+++ AOR/Model/NoteLine.cs
@@ -2,11 +2,11 @@
 {
     public class NoteLine
     {
-        public byte Tone;
+        public short Tone;
         public uint StartTime;
         public uint EndTime;
 
-        public NoteLine(byte tone, uint start, uint end)
+        public NoteLine(short tone, uint start, uint end)
         {
             Tone = tone;
             StartTime = start;
